(PUP-8066) Updated a set of acceptance tests for fips compliance
The tests have been adjusted to dynamically detect fips mode using a
fips fact so they can execute on both fips and non-fips platforms.

Updated versions of acceptance tests to run in FIPS mode:
 - acceptance/tests/face/loadable_from_modules.rb
 - acceptance/tests/modules/build/build_agent.rb
 - acceptance/tests/modules/build/build_should_not_create_changes.rb
 - acceptance/tests/modules/install/basic_install.rb
 - acceptance/tests/modules/install/with_version.rb
 - acceptance/tests/parser_functions/calling_all_functions.rb
 - acceptance/tests/puppet_apply_a_file_should_create_a_file_and_report_the_md5.rb
 - acceptance/tests/resource/file/content_attribute.rb
 - acceptance/tests/resource/file/source_attribute.rb
 - acceptance/tests/ticket_1334_clientbucket_corrupted.rb
 - acceptance/tests/ticket_4622_filebucket_diff_test.rb
 - acceptance/tests/ticket_6541_invalid_filebucket_files.rb

+++ acceptance/tests/modules/build/build_agent.rb
@@ -1,6 +1,12 @@
 test_name "puppet module build (agent)"
 
 agents.each do |agent|
+    
+  if (on(agent, facter("find in_fips_mode")).stdout =~ /true/)
+    puts "Module build, loading and installing not supported on fips enabled platforms"
+    next
+  end
+
   teardown do
     on agent, 'rm -rf bar'
   end
